mcp-server/pyproject.toml
[project]
name = "inthezon-mcp-server"
version = "0.1.0"
description = "MCP server for Inthezon Amazon data"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "questionary>=2.0.0",
    "rich>=13.0.0",
]

[project.scripts]
inthezon-mcp = "server:mcp.run"
inthezon-cli = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
